Move sending related methods to KafkaMessageSender
As describted by https://github.com/UST-MICO/mico/issues/762 we
want to split the receiving and the sending logic.
